+++ spec/models/early_career_payments/eligibility_answers_presenter_spec.rb
@@ -1,0 +1,21 @@
+require "rails_helper"
+
+RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
+  let(:eligibility_attributes) do
+    {
+      nqt_in_academic_year_after_itt: true
+    }
+  end
+  let(:eligibility) { claim.eligibility }
+  let(:claim) { build(:claim, eligibility: build(:early_career_payments_eligibility, eligibility_attributes)) }
+
+  subject(:presenter) { described_class.new(eligibility) }
+
+  it "returns an array of questions and answers to be presented to the user for checking" do
+    expected_answers = [
+      [I18n.t("early_career_payments.questions.nqt_in_academic_year_after_itt"), "Yes", "nqt-in-academic-year-after-itt"]
+    ]
+
+    expect(presenter.answers).to eq(expected_answers)
+  end
+end